+++ src/main/core.js
@@ -165,9 +165,15 @@ class Core {
       // '-connect=192.168.0.234',
       // //
     ];
+    //After core forksblocks clear out that field. 
     if (settings.forkBlocks) {
       params.push('-forkblocks=' + settings.forkBlocks);
       UpdateSettings({ forkBlocks: 0 });
+    }
+    // If WalletClean is set as true then run the walletclean command, core will auto start back up normaly after this is done
+    if (settings.walletClean) {
+      params.push('-walletclean');
+      UpdateSettings({walletClean: false});
     }
     //Avatar is default so only add it if it is off.
     if (!settings.avatarMode) {