+++ src/shared/lib/settings.js
@@ -65,6 +65,7 @@ export const defaultSettings = {
   windowX: undefined,
   windowY: undefined,
   forkBlocks: 0,
+  walletClean: false
 };
 
 export let tempSettings = null;